finished c implementation

+++ c_implementation/sigmoid.h
@@ -5,6 +5,6 @@
 #define EULER_NUMBER_F 2.71828182846
 #define EULER_NUMBER_L 2.71828182845904523536
 
-long double sigmoidl(long double n);
+ double sigmoidl( double n);
 
 #endif
